+++ Sources/Tokenizer/Layout/ConstraintModifier.swift
@@ -9,8 +9,8 @@
 import Foundation
 
 public enum ConstraintModifier {
-    case multiplied(by: Double)
-    case divided(by: Double)
-    case offset(by: Double)
-    case inset(by: Double)
+    case multiplied(by: Float)
+    case divided(by: Float)
+    case offset(by: Float)
+    case inset(by: Float)
 }
